+++ src/features/docBlocker/block/property.ts
@@ -1,0 +1,12 @@
+import { Block } from "../block";
+import { Doc, DocType } from "../doc";
+
+
+export default class Property extends Block {
+
+  protected pattern: RegExp = /^(\s*property)\s+/i;
+
+  public constructDoc(): Doc {
+    return new Doc(DocType.Property);
+  }
+}